API description for conference server drivers
Instead of relying on a hardcoded Big Blue Button server driver, the
plugin is now capable of providing different drivers for different
conference servers. The abstract API for accessing a conference server
is described in the \ElanEv\Driver\DriverInterface.

Basically, a server has to support to create new meeting which are
to be described by parameters (see the \ElanEv\Driver\MeetingParameter
class for the details). A meeting is then identified by a unique id.

Attendees can join a meeting by accessing a URL which is obtained from
the getJoinMeetingUrl() method of the driver implementation.

Additionally, the isMeetingRunning() method can be used to check if
a meeting with a particular id is already running.