Provide instructions for missing Opencast series
If recordings through Opencast are enabled but the Opencast Plugin
is inactive in the course, or the series is missing, the plugin
correctly rejects enabeling recordings.
However, it did not say how to create an Opencast Series correctly.
This year, lecturors joining our teaching corpus had serve issues
getting started.