Verify success of DFNConf meeting creation
DfnVc::createMeeting() now checks the return code of the API call
for sco-update/ type=meeting.

Otherwise it may happen that rooms are displayed in Stud.IP that
never have been created in the backend.
These rooms can't be deleted without touching the database because
deletion checks if the API call was successful.

Resolves #355