Make recording auto start default configurable
This commit introduces a new config option for the admin allowing to
specify if the recording of BBB meeting rooms should automatically
start if not yet set in the meeting room config.

In our case, the recording must not start prior to everybody in the
room providing their consent. If no consent for recording is established,
the seminar will take place without being recorded.