Disable auto-start button if recording unchecked
If recording checkbox in a room's settings is not checked,
the checkbox "allow webcam recording" already gets disabled
and unchecked.
This commit will apply the same behavior to the auto-start
recordings checkbox ("Aufzeichnung automatisch starten").

Note that this might be unwanted for some institutions
because it flips a default for new conference rooms:
Since recording is probably always disabled when creating
a new room, "Aufzeichnung automatisch starten" will now
be disalbed by default, too.

So please check with some institutions before merging.